🚚 move: reubica controlador y request de archivos médicos al namespace Files y actualiza routes/api.php
