+++ app/Http/Requests/Files/StoreMedicalRecordFileRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Files;
 
 use App\Http\Requests\BaseRequest;
